Swanky: Test batterycutoff returned value
Do not shutdown DUT and show warning if batterycutoff fails.

BUG=chrome-os-partner:29077
TEST=Factory wipe and check the battery is cut-off.
     Factory reset and check the battery is cut-off.

Reviewed-on: https://chromium-review.googlesource.com/205384
Tested-by: JAMES C Chen <james.c.chen@intel.com>
Commit-Queue: JAMES C Chen <james.c.chen@intel.com>

+++ overlay-swanky/chromeos-base/chromeos-bsp-swanky/files/battery_cut_off.sh
@@ -35,9 +35,12 @@ else
   echo "==== Cutting off battery. Wait 10 seconds. ====" >"$TTY"
   echo "===============================================" >"$TTY"
 fi
-ectool batterycutoff
-shutdown -h now
-sleep 15
+
+if ectool batterycutoff; then
+  sleep 2
+  shutdown -h now
+  sleep 15
+fi
 
 # Couldn't have reached here
 if [ -e "${IMG_PATH}/cutoff_failed.png" ]; then
